requirements.txt
psycopg==3.1.12
psycopg-binary==3.1.12
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
